some func restored

+++ colour_filter_utils.hpp
@@ -33,7 +33,7 @@ uint16_t ffv1(uint16_t L,uint16_t T,uint16_t TL){
 	return max - (TL - min);
 }
 
-uint8_t i_clamp(int16_t a,int16_t lower,int16_t upper){
+uint16_t i_clamp(int16_t a,int16_t lower,int16_t upper){
 	if(a < lower){
 		return lower;
 	}
